added test to query for guid and did some refactoring
